+++ soft/main/utils/matrix.h
@@ -11,15 +11,15 @@ class Matrix
 
     unsigned int            _rows;
     unsigned int            _columns;
-    vector<vector<double> >  _data;
+    vector<vector<float> >  _data;
 
   public:
 
-    Matrix(unsigned int, unsigned int, double initial = 0.0);
+    Matrix(unsigned int, unsigned int, float initial = 0.0);
 
-    vector<vector<double> > data();
-    void set_data(vector<vector<double> >);
-    void set(unsigned int, unsigned int, double);
+    vector<vector<float> > data();
+    void set_data(vector<vector<float> >);
+    void set(unsigned int, unsigned int, float);
 
     unsigned int rows(void);
     unsigned int columns(void);
@@ -31,19 +31,19 @@ class Matrix
     Matrix minor(int i);
     Matrix cofactor(int row, int column);
   
-    double determinate(void);
+    float determinate(void);
 
     void display(string name);
 
-    double operator()(const unsigned &, const unsigned &);
+    float operator()(const unsigned &, const unsigned &);
     Matrix operator+(Matrix);
     Matrix operator-(Matrix);
     Matrix operator*(Matrix);
     Matrix operator%(Matrix);
-    Matrix operator+(double);
-    Matrix operator-(double);
-    Matrix operator*(double);
-    Matrix operator/(double);
+    Matrix operator+(float);
+    Matrix operator-(float);
+    Matrix operator*(float);
+    Matrix operator/(float);
 
     static Matrix identity(int n);
